Fix bugs
For e.g. N=100 the blocks scheme was failing before.

The work balacing was fixed by rounding the blockSize up while also making sure
that it is a multiple of 64.

The counting of primes was also fixed by excluding half_size from the range.
For consistency half_size was also removed from the ranges processed by the
kernels.

Fix counting